Time: 1 ms (86.84%) | Memory: 41.2 MB (29.28%) - LeetSync

+++ 69-sqrtx/sqrtx.java
@@ -6,7 +6,7 @@ class Solution {
         while(l<=r){
             int  mid = l+ (r-l)/2;
 
-            if (mid * mid == x){
+            if ((long)mid * mid ==(long) x){
                 return  mid;
             }
             else if((long)mid * mid >(long) x){
